Added AverageUtil and Null input behaviour into Utils. Readability changes.

+++ src/main/java/ee/taltech/calculator/dto/CalculationResult.java
@@ -9,7 +9,7 @@ import java.util.List;
 @Setter
 public class CalculationResult {
     private Integer min;
-    private Float averageOfPositives;
+    private Double averageOfPositives;
     private List<Integer> odds;
     private Integer sumOfOdds;
     private List<Integer> powerOfFour;
